+++ scriptconfig/__init__.py
@@ -145,7 +145,7 @@ Ignore:
     mkinit ~/code/scriptconfig/scriptconfig/__init__.py --nomods --relative -w
 """
 
-__version__ = '0.8.1'
+__version__ = '0.8.2'
 
 __submodules__ = {
     'modal': None,